backend/requirements.txt
# Backend Requirements
# Install with: pip install -r requirements.txt

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# File upload support - CRITICAL
python-multipart>=0.0.6

# OpenAI (for Whisper STT and future GPT usage)
openai>=1.0.0

# Anthropic (for Claude)
anthropic>=0.7.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL driver

# Alternative: MongoDB
# motor>=3.3.0
# pymongo>=4.5.0

# Vector database (FAISS for RAG)
faiss-cpu>=1.7.4
# OR for GPU support:
# faiss-gpu>=1.7.4

# Embeddings and NLP
sentence-transformers>=2.2.0
tiktoken>=0.5.0  # Token counting for LLMs

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Clerk authentication
pyclerk>=0.0.5

# WebSocket support
websockets>=12.0

# Utilities
python-jose[cryptography]>=3.3.0  # JWT handling
passlib[bcrypt]>=1.7.4  # Password hashing (if needed)
python-dateutil>=2.8.2

# Logging and monitoring
python-json-logger>=2.0.7

# Audio processing (optional, for format conversion)
# pydub>=0.25.1
# ffmpeg-python>=0.2.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For TestClient

# Code quality (development)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0